Disable CPU1 during flash write.
Looking at an old project I see this was done and remember debugging
the issue in the past.  It makes sence that writing flash needs
a very quiet environment to work properly.

With this, I likely don't need the debug stuff I added earlier, but
I'll keep in it place until I see eveything working as expected.